+++ source/rafcon/core/states/state.py
@@ -849,10 +849,11 @@ class State(Observable, YAMLObject, JSONObject, Hashable):
         :return:
         """
         if state_id is None:
-            state_id = state_id_generator()
+            state_id = state_id_generator(used_state_ids=[self.state_id])
         if not self.is_root_state and not self.is_root_state_of_library:
-            while state_id in self.parent.states:
-                state_id = state_id_generator()
+            used_ids = self.parent.states.keys() + [self.parent.state_id, self.state_id]
+            if state_id in used_ids:
+                state_id = state_id_generator(used_state_ids=used_ids)
 
         self._state_id = state_id