+++ source/rafcon/gui/controllers/menu_bar.py
@@ -31,11 +31,9 @@ from rafcon.core.singleton import state_machine_manager, library_manager
 from rafcon.core.states.barrier_concurrency_state import BarrierConcurrencyState
 from rafcon.core.states.preemptive_concurrency_state import PreemptiveConcurrencyState
 from rafcon.gui import singleton as gui_singletons
-import rafcon.gui.helpers.label as gui_helper_label
 from rafcon.gui.config import global_gui_config
 from rafcon.gui.controllers.preferences_window import PreferencesWindowController
 from rafcon.gui.controllers.utils.extended_controller import ExtendedController
-import rafcon.gui.helpers.state_machine as gui_helper_state_machine
 from rafcon.gui.models.abstract_state import AbstractStateModel
 from rafcon.gui.runtime_config import global_runtime_config
 from rafcon.gui.utils import constants
@@ -44,6 +42,11 @@ from rafcon.gui.views.preferences_window import PreferencesWindowView
 from rafcon.gui.views.main_window import MainWindowView
 from rafcon.gui.views.utils.about_dialog import AboutDialogView
 import rafcon.gui.backup.session as backup_session
+
+import rafcon.gui.helpers.label as gui_helper_label
+import rafcon.gui.helpers.state_machine as gui_helper_state_machine
+import rafcon.gui.helpers.utility as gui_helper_utility
+
 from rafcon.utils import plugins
 from rafcon.utils import log
 
@@ -89,6 +92,16 @@ class MenuBarController(ExtendedController):
     def destroy(self):
         super(MenuBarController, self).destroy()
         self.full_screen_window.destroy()
+
+    @staticmethod
+    def create_logger_warning_if_shortcuts_are_overwritten_by_menu_bar():
+        shortcut_dict = global_gui_config.get_config_value('SHORTCUTS')
+        shortcut_key_patterns = [elem for l in shortcut_dict.values() for elem in l]
+        for key in ['E', 'F', 'V', 'X', 'H', 'e', 'f', 'v', 'x', 'h']:
+            if '<Alt>' + key in shortcut_key_patterns:
+                dict_pair = {k: v for k, v_list in shortcut_dict.iteritems() for v in v_list if '<Alt>' + key == v}
+                logger.warning("Your current shortcut {0} is not working because a menu-bar access key "
+                               "is overwriting it.".format(dict_pair))
 
     def register_view(self, view):
         """Called when the View was registered"""
@@ -157,6 +170,8 @@ class MenuBarController(ExtendedController):
         self.view['menu_edit'].connect('select', self.check_edit_menu_items_status)
         self.registered_view = True
         self._update_recently_opened_state_machines()
+        # do not move next line - here to show warning in GUI debug console
+        self.create_logger_warning_if_shortcuts_are_overwritten_by_menu_bar()
 
     @ExtendedController.observe('config', after=True)
     def on_config_value_changed(self, config_m, prop_name, info):
@@ -307,6 +322,12 @@ class MenuBarController(ExtendedController):
 
         self.add_callback_to_shortcut_manager('is_start_state', partial(self.call_action_callback,
                                                                         "on_toggle_is_start_state_active"))
+        callback_function = partial(self.call_action_callback, "on_add_transitions_from_closest_sibling_state_active")
+        self.add_callback_to_shortcut_manager('transition_from_closest_sibling_state', callback_function)
+        callback_function = partial(self.call_action_callback, "on_add_transitions_to_closest_sibling_state_active")
+        self.add_callback_to_shortcut_manager('transition_to_closest_sibling_state', callback_function)
+        callback_function = partial(self.call_action_callback, "on_add_transitions_to_parent_state_active")
+        self.add_callback_to_shortcut_manager('transition_to_parent_state', callback_function)
         self.add_callback_to_shortcut_manager('group', partial(self.call_action_callback, "on_group_states_activate"))
         self.add_callback_to_shortcut_manager('ungroup', partial(self.call_action_callback,
                                                                  "on_ungroup_state_activate"))
@@ -460,6 +481,7 @@ class MenuBarController(ExtendedController):
         for item_name, shortcuts in global_gui_config.get_config_value('SHORTCUTS', {}).iteritems():
             if shortcuts and item_name in self.view.buttons:
                 self.view.set_menu_item_accelerator(item_name, shortcuts[0])
+        self.create_logger_warning_if_shortcuts_are_overwritten_by_menu_bar()
 
     def on_delete_check_sm_modified(self):
         if state_machine_manager.has_dirty_state_machine():
@@ -517,6 +539,18 @@ class MenuBarController(ExtendedController):
     @staticmethod
     def on_toggle_is_start_state_active(widget, data=None):
         return gui_helper_state_machine.selected_state_toggle_is_start_state()
+
+    @staticmethod
+    def on_add_transitions_from_closest_sibling_state_active(widget, data=None):
+        return gui_helper_utility.add_transitions_from_closest_sibling_state_to_selected_state()
+
+    @staticmethod
+    def on_add_transitions_to_closest_sibling_state_active(widget, data=None):
+        return gui_helper_utility.add_transitions_to_closest_sibling_state_from_selected_state()
+
+    @staticmethod
+    def on_add_transitions_to_parent_state_active(widget, data=None):
+        return gui_helper_utility.add_transitions_from_selected_state_to_parent()
 
     def on_copy_selection_activate(self, widget, data=None):
         self.shortcut_manager.trigger_action("copy", None, None)