+++ source/rafcon/gui/views/tool_bar.py
@@ -18,7 +18,6 @@ from gtkmvc import View
 from rafcon.gui import glade
 from rafcon.gui.helpers.label import create_label_widget_with_icon
 from rafcon.gui.utils import constants
-from rafcon.utils.i18n import _
 
 
 class ToolBarView(View):