+++ tests/assets/unit_test_state_machines/stepping_test_with_library/stepping_test_with_library_GLSUJY/Decimate_bottles_NDIVLD/script.py
@@ -1,0 +1,6 @@
+import time
+
+def execute(self, inputs, outputs, gvm):
+    outputs['bottles'] = inputs['bottles'] - 1
+    time.sleep(0.01)
+    return 0